requirements.txt
# DS project's dependencies that must be deployed to the scrapy cloud

kafka-python==0.9.4
boto==2.38.0
requests==2.5.3
msgpack-python==0.4.2
sqlitedict==1.3.0
SQLAlchemy==1.0.5
jq==0.1.2
PyYAML==3.11
psycopg2==2.6.1
-e git+git@github.com:scrapinghub/collection-scanner.git#egg=collection_scanner
-e git+git@github.com:scrapinghub/kafka-scanner.git#egg=kafka_scanner
hubstorage==0.16.1
mysql-python==1.2.5

# pysftp and it's deps
pysftp==0.2.8
ecdsa==0.13
paramiko==1.15.2
pycrypto==2.6.1
wsgiref==0.1.2

# retrying and it's deps
retrying==1.2.3
wheel==0.24.0


# tests
mock==1.0.1
moto==0.4.2
coverage==3.7.1
nose==1.3.6


